Removed player health text
