+++ Assets/Scripts/GameTextManager.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: ea98ed9dc76e3435fa8ac46146bc90f2
+guid: 01e526ec3613349e59e5832088f41bb0
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
